Clean up of solr functions, addition of solr layerdate, removal of extra fields and general solr reorganisation
