+++ hypermap/aggregator/utils.py
@@ -1,16 +1,9 @@
-import pysolr
-import requests
 import urllib2
 import re
-import logging
-import time
-import math
-import json
 import sys
+import math
 import traceback
 
-from django.conf import settings
-from urlparse import urlparse
 
 from owslib.wms import WebMapService
 from owslib.tms import TileMapService
@@ -172,221 +165,3 @@ def get_sanitized_endpoint(url):
         match = re.search(esri_string, sanitized_url)
         sanitized_url = url[0:(match.start(0)+len(esri_string))]
     return sanitized_url
-
-
-class OGP_utils(object):
-
-    @staticmethod
-    def good_coords(coords):
-        """ passed a string array """
-        if (len(coords) != 4):
-            return False
-        for coord in coords[0:3]:
-            try:
-                num = float(coord)
-                if (math.isnan(num)):
-                    return False
-                if (math.isinf(num)):
-                    return False
-            except ValueError:
-                return False
-        return True
-
-    solr_url = getattr(settings, 'SOLR_URL', 'http://localhost:8983/solr/geonode24')
-    solr = pysolr.Solr(solr_url, timeout=60)
-    logger = logging.getLogger("hypermap.aggregator.utils")
-
-    @staticmethod
-    def get_domain(url):
-        urlParts = urlparse(url)
-        hostname = urlParts.hostname
-        if hostname == "localhost":
-            return "Harvard"  # assumption
-        return hostname
-
-    @staticmethod
-    def is_solr_up():
-        solr_url = getattr(settings, 'SOLR_URL', 'http://localhost:8983/solr/geonode24')
-        solr_url_parts = solr_url.split('/')
-        admin_url = '/'.join(solr_url_parts[:-1]) + '/admin/cores'
-        params = {'action': 'STATUS', 'wt': 'json'}
-        try:
-            req = requests.get(admin_url, params=params)
-            response = json.loads(req.text)
-            status = response['status']
-            if status:
-                response = True
-        except requests.exceptions.RequestException:
-            response = False
-        return response
-
-    @staticmethod
-    def layer_to_solr(layer, i=0):
-        try:
-            bbox = [float(layer.bbox_x0), float(layer.bbox_y0), float(layer.bbox_x1), float(layer.bbox_y1)]
-            for proj in layer.srs.values():
-                if '102113' == proj['code']:
-                    bbox = mercator_to_llbbox(bbox)
-            # to do need to figure out it hypermap will only be dealing with remote servives
-            storeType = "remoteStore"
-            # We will need to get the layer date lists if they exist here and use this instead as a date range
-            date = layer.created
-            if (OGP_utils.good_coords(bbox)) is False:
-                print 'no coords in layer ', layer.title
-                return
-            if (OGP_utils.good_coords(bbox)):
-                print 'in utils.layer_to_solr, bbox = ', bbox
-                username = ""
-                minX = bbox[0]
-                minY = bbox[1]
-                maxX = bbox[2]
-                maxY = bbox[3]
-                if (minY > maxY):
-                    tmp = minY
-                    minY = maxY
-                    maxY = tmp
-                if (minX > maxX):
-                    tmp = minX
-                    minX = maxX
-                    maxX = tmp
-                centerY = (maxY + minY) / 2.0
-                centerX = (maxX + minX) / 2.0
-                halfWidth = (maxX - minX) / 2.0
-                halfHeight = (maxY - minY) / 2.0
-                area = (halfWidth * 2) * (halfHeight * 2)
-                if (minX < -180):
-                    minX = -180
-                if (maxX > 180):
-                    maxX = 180
-                if (minY < -90):
-                    minY = -90
-                if (maxY > 90):
-                    maxY = 90
-                # ENVELOPE(minX, maxX, maxY, minY) per https://github.com/spatial4j/spatial4j/issues/36
-                wkt = "ENVELOPE({:f},{:f},{:f},{:f})".format(minX, maxX, maxY, minY)
-                # default for now as we do not have capabilitty to resolve vector/raster layers
-                dataType = "Polygon"
-                institution = "Harvard"
-                servicetype = layer.service.type
-                owsUrl = layer.service.url
-                if storeType == "remoteStore":
-                    institution = "Remote"
-                    if servicetype == "ESRI":
-                        dataType = "RESTServices"
-                    else:
-                        dataType = "WMSServices"
-                domain = OGP_utils.get_domain(owsUrl)
-                if (i == 0):
-                    i = layer.title
-                # we will need to resolve the projection issue at a later date
-                OGP_utils.solr.add([{
-                                    "LayerId": "HyperMapLayer_" + str(i),
-                                    "Name": layer.title,
-                                    "LayerDisplayName": layer.title,
-                                    "Institution": institution,
-                                    "Publisher": username,
-                                    "Originator": domain,
-                                    "ServiceType": servicetype,
-                                    "ContentDate": date,
-                                    "Access": "Public",
-                                    "DataType": dataType,
-                                    "Availability": "Online",
-                                    "Location": '{"layerInfoPage": "' + layer.get_absolute_url() + '"}',
-                                    "Abstract": layer.abstract,
-                                    "SrsProjectionCode": 'EPSG:4326',
-                                    "MinY": minY,
-                                    "MinX": minX,
-                                    "MaxY": maxY,
-                                    "MaxX": maxX,
-                                    "CenterY": centerY,
-                                    "CenterX": centerX,
-                                    "HalfWidth": halfWidth,
-                                    "HalfHeight": halfHeight,
-                                    "Area": area,
-                                    "bbox_rpt": wkt}])
-                OGP_utils.logger.error("solr record saved: " + layer.title)
-        except Exception as e:
-            if e.message.startswith("Connection") or e.message.startswith("[Reason: java.lang.OutOfMemoryError:"):
-                OGP_utils.solr.add([{
-                                     "LayerId": "HarvardWorldMapLayer_" + str(i),
-                                     "Name": layer.title,
-                                     "LayerDisplayName": layer.title,
-                                     "Institution": institution,
-                                     "Publisher": username,
-                                     "Originator": domain,
-                                     "ServiceType": servicetype,
-                                     "ContentDate": date,
-                                     "Access": "Public",
-                                     "DataType": dataType,
-                                     "Availability": "Online",
-                                     "Location": '{"layerInfoPage": "' + layer.get_absolute_url() + '"}',
-                                     "Abstract": layer.abstract,
-                                     "SrsProjectionCode": 'EPSG:4326',
-                                     "MinY": minY,
-                                     "MinX": minX,
-                                     "MaxY": maxY,
-                                     "MaxX": maxX,
-                                     "CenterY": centerY,
-                                     "CenterX": centerX,
-                                     "HalfWidth": halfWidth,
-                                     "HalfHeight": halfHeight,
-                                     "Area": area,
-                                 }])
-                OGP_utils.logger.error("failed solr record saved after retry: " + layer.title)
-            else:
-                OGP_utils.logger.error("error in layer_to_solr processing layer: " + e.message)
-
-    @staticmethod
-    def clear_solr():
-        """Clear all indexes in the solr core"""
-        OGP_utils.solr.delete(q='*:*')
-        print 'Solr core cleared'
-
-    @staticmethod
-    def solr_to_solr():
-        """create Solr records for layers in another Solr OGP instance"""
-        ogp_solr = pysolr.Solr('http://geodata.tufts.edu/solr/')
-        wm_solr = pysolr.Solr('http://localhost:8983/solr/all', timeout=60)
-        count = 3200
-        while (True):
-            docs = ogp_solr.search("*:*", start=count)
-            count += len(docs)
-            if (len(docs) == 0):
-                return
-            time.sleep(1)
-            print "count = ", count
-            for doc in docs:
-                minX = doc['MinX']
-                minY = doc['MinY']
-                maxX = doc['MaxX']
-                maxY = doc['MaxY']
-                wkt = "ENVELOPE({:f},{:f},{:f},{:f})".format(minX, maxX, maxY, minY)
-                if (-90 <= minY <= 90 and -90 <= maxY <= 90 and -180 <= minX <= 180 and -180 <= maxX <= 180):
-                    try:
-                        wm_solr.add([{
-                            "LayerId": doc['LayerId'],
-                            "Name": doc['Name'],
-                            "LayerDisplayName": doc['LayerDisplayName'],
-                            "Institution": "OGP-" + doc['Institution'],  # "Harvard",
-                            "Publisher": doc['Publisher'],
-                            "Originator": doc['Originator'],
-                            "Access": 'Public',
-                            "DataType": doc['DataType'],
-                            "ContentDate": doc['ContentDate'],
-                            "Availability": doc['Availability'],
-                            "Location": doc['Location'],
-                            "Abstract": doc['Abstract'],
-                            "MinY": doc['MinY'],
-                            "MinX": doc['MinX'],
-                            "MaxY": doc['MaxY'],
-                            "MaxX": doc['MaxX'],
-                            "CenterY": doc['CenterY'],
-                            "CenterX": doc['CenterX'],
-                            "HalfWidth": doc['HalfWidth'],
-                            "HalfHeight": doc['HalfHeight'],
-                            "Area": doc['Area'],
-                            "bbox_rpt": wkt}])
-                    except KeyError as e:
-                        print doc['LayerDisplayName', e]
-                else:
-                    print "bad bounds: ", doc
